requirements.txt
first
pyee
aiohttp<4.0
slackclient>2.0
